refactor(web): 코드 리뷰 반영 — stats 분리, 미사용 import 제거
- MonthlyHeatmap을 별도 파일로 분리 (routine-stats.tsx 254줄 → 147줄)
- routine-page.tsx 미사용 RoutineTemplateRow import 제거
- use-routines.ts fetchStats에 401 처리 추가
